se reduce prop drilling en turnoScreen
